Add files via upload
GL/glew.h and GLFW/glfw3.h libraries are used. Extra interactive aspects of this project are ball changes its color to red when it touches the redline area. You can  move the red line by using up and down arrows in keyboard. You can increase or decrease the speed of the ball on the screen by pressing left or right arrows on keyboard but this portion of code is in comment block to if you want to use it don't forget to make it active